Extract pretender code to its own module for future move to another package
