+++ lib/index.js
@@ -3,7 +3,7 @@ import IdentityManager from "./identity-manager";
 import association from "./association";
 import trait from "./trait";
 import Response from "./response";
-import Server, { createServer, defaultPassthroughs } from "./server";
+import Server, { createServer } from "./server";
 import Model from "./orm/model";
 import Collection from "./orm/collection";
 import Serializer from "./serializer";
@@ -47,6 +47,8 @@ import _utilsIsAssociation from "./utils/is-association";
 import _utilsReferenceSort from "./utils/reference-sort";
 import _utilsUuid from "./utils/uuid";
 
+import PretenderInterceptor from "./mock-server/pretender-config";
+
 /**
   @hide
 */
@@ -72,7 +74,6 @@ export {
   RestSerializer,
   hasMany,
   belongsTo,
-  defaultPassthroughs,
   createServer,
   Server,
   Factory,
@@ -105,6 +106,7 @@ export {
   _utilsIsAssociation,
   _utilsReferenceSort,
   _utilsUuid,
+  PretenderInterceptor,
 };
 
 export default {
